Fix invalid period in a wallet notice text
Summary:
According to one test assertion on 2021-03-05 we expect "... will last until about 2021-04-02 (4 weeks)."
This test was failing occasionally.

I found that one of the dates we were comparing was N microseconds ahead.
So, the period was less than a second shorter. Carbon didn't round this.

A solution is to just use the same time on both compared dates.

Test Plan: `./phpunit`

Reviewers: #hkccp_developers, vanmeeuwen

Reviewed By: #hkccp_developers, vanmeeuwen

Subscribers: #hkccp_developers

Differential Revision: https://git.kolab.org/D2365