Laravel Passport support
Summary:
Replaces the use of tymon:jwt with laravel:passport.

A password grant client is used create tokens for the webclient in the
same fashion as we used to the tymon:jwt solution.
The same password grant client can be used for other client applications.

Notes:
* We're not currently purging invalidated/expired tokens. This can be
done via artisan command or scheduled task from php.

Differential Revision: https://git.kolab.org/D2494

* Pass the refresh token
* Custom authentication hook for passport
* Added 2fa to user verification
* Rely on second factor authentication in user model
* Disabled unnecessary passport routes
* Fixed password-reset