Laravel Passport support
Summary:
Replaces the use of tymon:jwt with laravel:passport.

A password grant client is used create tokens for the webclient in the
same fashion as we used to the tymon:jwt solution.
The same password grant client can be used for other client applications.

Notes:
* We're not currently purging invalidated/expired tokens. This can be
done via artisan command or scheduled task from php.

Reviewers: #hkccp_developers, machniak, vanmeeuwen

Reviewed By: #hkccp_developers, machniak, vanmeeuwen

Subscribers: vanmeeuwen, alec

Differential Revision: https://git.kolab.org/D2494