Get meet to work over the kolab.io proxy
The api has moved to /meetmedia/api (to not clash with /meet/ that we
use for kolab4 already.
The websocket for signaling is at /meetmedia/signaling.

The websocket is now a wss/https socket, because ws/http just silently
fails in the browser (some security mechanism?).

For nginx proxying the following is required (both pointing to
127.0.0.1:12443 on the local system)

location /meetmedia/api {
	proxy_pass http://127.0.0.1:18006;
	proxy_redirect   off;
	proxy_set_header Host $host;
	proxy_set_header X-Real-IP $remote_addr;
	proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
	proxy_set_header X-Forwarded-Host $host;
	proxy_set_header X-Forwarded-Proto $scheme;
	proxy_no_cache 1;
	proxy_cache_bypass 1;
}

location /meetmedia {
	proxy_pass http://127.0.0.1:18006;
	proxy_http_version 1.1;
	proxy_set_header Upgrade $http_upgrade;
	proxy_set_header Connection 'upgrade';
	proxy_set_header Host $host;
}