Removed returning peer code
A jwt was used to store in the session that a peer has access to the
room. So a returning user could bypass access control and lobby.
Finally the client would get a different signal without the turn
servers.

It seems we don't require this since we do access control in kolab4, so
returning is not much different to just connecting to a room.