Use time limited TURN credentials
Instead of adding random username/password combinations to a redis db,
that then has to be maintained somehow, we can use the mechanism for
time-limited turn credentials that are based on a static secret.

Instead of a completely static secret it would be possible to store a
secret in the redis db, that could be rotated on an interval.