We have to send signals I'm relatiely sure, so remove TODO
AFAIK mediasoup will not send any signal to the client, this must be
done by our implementation.
The observer signal on the client side is triggered by our own call to
e.g. client.pause() and is primarily useful other parts of the
application, but doesn't save us the signaling.

See also https://mediasoup.discourse.group/t/consumer-observer-on-pause-fn-event-is-not-emitting/2216/3