Use domain_root_dn() for all kind of objects consistently
Summary:
It might be not needed for groups/shared folders/resources as we
create them in custom domains only, but I think it's better to treat them all
the same as users.

It will have small performance impact, but the result of this
method is internally cached by Net/LDAP3. If this becomes a problem we can
extend it with Redis support and use this cache (it already supports Memcache).

Test Plan: `./phpunit`

Reviewers: #hkccp_developers, vanmeeuwen

Reviewed By: #hkccp_developers, vanmeeuwen

Subscribers: #hkccp_developers

Differential Revision: https://git.kolab.org/D3155