Set all require environment variables for the kolab container
Summary:
and avoid relying on default values in environment variables.

We need those variables to be set, and it becomes a lot clearer if we fail to
if we don't just silently use a fallback value.

Also: Create and set mysql username/password for localhost and '%'
(the localhost user is not accessible over 127.0.0.1)

Differential Revision: https://git.kolab.org/D3268