Display Subscriptions section for all entitleable objects
Summary:
Display Subscriptions in the UI for editing, but also creating entitleable objects ...except where we display Packages.

The main idea is to display the expected price for the new object, but also give a better overview on what the user pays for. We already did it for domains, now for all entitleables.

Test Plan: `./phpunit`

Reviewers: #hkccp_developers, mollekopf

Reviewed By: #hkccp_developers, mollekopf

Subscribers: mollekopf, #hkccp_developers

Differential Revision: https://git.kolab.org/D3614