Run all dns queries over pdns
By using a recursor + the pdns-sql we can resolve all queries.
The ip:port syntax is not in resolv.conf is not actually supported by
libc (but seems to work from php), so e.g. dig won't work.
I think the solution will be to not use host-networking so we can just
specify pdns as dns server.