An example standalone ds389 container
Based on the upstream idea how to run one using /usr/libexec/dirsrv/dscontainer.
See also:
* https://directory.fedoraproject.org/docs/389ds/howto/howto-deploy-389ds-on-openshift.html
* https://fy.blackhats.net.au/blog/html/2020/03/28/389ds_in_containers.html

This works fine for running a container. However, I have struggled for a
bit and eventually given up to initialize it.

There does not seem to be a decent way to to do it from within the
container (maybe there is a way to prepare the db without a running ldap
with some of the bak2db tools, I don't know.)

The alternative seems to be doing it from the kolab container (or a
dedicated ldap-init container, but that's not really a thing in
docker-compose world), but the current setup procedure (pykolab) relies on e.g.
dscreate, and the procedures in the above links rely on running commands
in the container, which we can't from the kolab container.
So unless we can actually do everything we need via the likes of ldapadd
etc, I don't know how this setup can be decently integrated into
docker-compose.

So I've given up on this endavour and decided we can live with ldap
being part of the kolab container until we eventually no longer require
it.