Only create the necessary minimum of skus for the config.prod
Summary:
In config.prod we just create the skus that we require, for the
domain-hosting and beta sku, we adjust the code so we enable the features if
the sku doesn't exist at all.

All sku related migrations have to be dropped from the main migrations,
as the skus are environment specific. I also dropped them from
config.demo, since we already set the appropriate values in the seeder,
so the migrations are pointless.

Original commits:
* Enable beta entitlements if the beta sku doesn't exist
* Drop sku related migrations
* ImapAdminSeeder
* Enable domain-hosting features if there is no domain-hosting sku
* Minimal set of skus in the admin seeder

Differential Revision: https://git.kolab.org/D4142