Example kustomize config
This is an example deployment of kolab on openshift.
It creates a kolab4 demo deployment with config.prod, including the
required mysql and redis dependencies, but no other dependencies yet (no
roundcube/imap/ldap/....).

The overlay deployment is an example how an external deployment config can be
applied (see the Makefile how the sources are uploaded). For the
config.prod deployment the overlay deployment is not required.