Move the .s2i scripts to the root, so the configure step works
Previously we only mounted the src/ directory into the container,
which means we can't apply config.prod (or alike).
This didn't matter for deployments with an external theme.

With this change we can deploy config.prod via the s2i scripts

The resulting container remains the same, so this still works with the
external overlays as they are.