Support for the kolab/horizon roles in the webapp container
In kubernetes we're splitting up the frontend(kolab) and the
backend(horizon). For initialization we want to make sure that only one
pod is running the initialization, and probably wait with fully starting
the frontend until the initialization is complete.

This would probably be better solved with an init container, but this is
a start.