.well-known/caldav|carddav redirect via nginx to https directly
Because we are behind a proxy we used to first redirect to http, and
then again to https, which I think thunderbird doesn't like.
Also, instead of redirecting in the nginx proxy we handle all redirects
in apache now, so we hit the same case in the docker and kubernetes
deployments.