podman based testexecution via testctl
We use a podman pod instead of docker-compose to run the tests for
better isolation, and for less interaction with the an existing
environment. This should also make integration into a CI system easier.

Some adjustmenets are required because everything connects over
localhost in a pod.

This patch also adds:
* A mechanism to apply an overlay during startup (used to apply the
  test overlay config.demo)
* Fixed the roundcube update from source code