Don't import during container start
Until very recently we only ran the openexchange import,
and this should probably be done on a schedule (e.g. daily) or during
seeding.

Also, this is now relatively slow, and we really don't want to block
every container start by downloading a bunch of stuff and running a ton
of queries.