Example content filter
This is using the pipe + script approach.
The benefit is that postfix takes care of spawning the script as
required.
The downside is performance and reliability according to the manual.

We currently execute amavis first and then the content filter.

Alternatives are:
* Script + spawn: Postfix manages the instances,
the script communicates via stdin/stdout (using smtp?).
Couldn't get it to work so far.
Performance and reliability potentially better according to the manual.
* Completely separate process that communicates via smtp. Same as amavis/wallace.
I'd like to avoid having to manage a separate process, but this would
definitely work.