Support for more item classes via hierarchical handling of EWS classes
Various subclasses to e.g. IPM.Note exist, but it seems we can just
handle them in a uniform way for migration.
We are not curorently using the TYPE constant of the item classes,
and with this change there are potentially multiple per class,
so I just left them for documentation.

This adds support for:
* IPM.Note.Rules.OofTemplate.Microsoft
* IPM.Schedule.Meeting.Request
* IPM.Schedule.Meeting.Canceled
* IPM.Note.SMIME.MultipartSigned
* REPORT.IPM.Note.NDR

Everything is just imported as regular mail:
* invitations contain ics
files
* SMIME contains the signature attachment
* NDR renders as a mail
* OOFTemplate also just renders as a mail