kolab_policy_ratelimit fix
After printing an action we should exit. Otherwise we trigger an
exception below and end up deferring instead of the dunno.

We do trigger this codepath every now and then (noticed due to
monitoring test messages), but I'm not sure why, or why we have to check for
the protocol state at all.