Disable the mandate if we retrieve an invalid one
Summary:
Right now we just retry to retrieve the mandate, which will then fail
again. I'm not sure if there is a "self-healing" case where this
behaviour makes sense. Also, we may want to send an email to the user
when this happens (the paymentfailure one looks like it could match this
case).

Differential Revision: https://git.kolab.org/D5028