Init command
Summary:
The init command is intended as an additional mechanism to migrations
and seeds.

The purpose of the init command is to create/update database entries
that are required.
It is meant to be executed after migration and seeder during startup,
or at runtime (e.g. when the configuration changed).
It should normally always be executed during startup.

Entries created in the init command do not have to be in
migrations/seeders additionally.

Differential Revision: https://git.kolab.org/D5068