Authentication caching for passwords
Summary:
This gets the time for authentication from 160ms to 0.5ms.

If somebody would get a hold of the cacheId,
a dictionary attack on the password becomes feasible (because the hash function is fast).
We only store this value on the internal redis instance, so the cacheId
could only be accessed if there already is breach, and the alternative would be to lower the bcrypt rounds to 6, which results
in 4ms, but that affects data at rest.

In the current implementation we incur one slow request every 60s,
and thus would also discover e.g. a password change after 60s.

Differential Revision: https://git.kolab.org/D5090