Shallow clone git repository
This results in a massive speedup for roundcube builds because we only
fetch 4.2 MB instead of 68MB of data for the roundcube repository alone.

We really only want the files and we remove the .git later on, so there
isn't really a downside. Unfortunately it's only possible to fetch
commits that are currently "advertised", which means they are currently
at the tip of a branch/tag, which is why we fall back to a full clone.
Fortunately we typically do clone the tip of a branch.