Use incrementing retry interval (backoff) in all jobs
Summary:
It should be always better to wait some time on error, instead of trying again immediately.
It gives more time to the job in case of some temporary error. It should work better
with short breaks, but if DB/IMAP/LDAP is dead for longer the job will eventually fail.

Also, unified number of attempts.

Reviewers: #hkccp_developers, mollekopf

Reviewed By: #hkccp_developers, mollekopf

Subscribers: #hkccp_developers

Differential Revision: https://git.kolab.org/D5108