Fold rate-limiting into the submission policy and get mailfilter to work
Summary:
We keep the existing separate rate-limiting endpoint for existing
installations for now. The new policy/submission endpoint can handle
rate-limiting as well for new installations

The postfix configuration now contains the necessary minimum with
regards to the policy hooks. We need the extra hooks to collect
recipients.

To support delegation, the reject_sender_login_mismatch check has been
removed, so we rely on the access policy to check sender address
ownership.

Some adjustments were required to get the mailfilter to function again,
such as including a sender in the request.
The mailfilter now accepts emails that are to non-local recipients or
have multiple recipents, to not block outgoing email (it's invoked for
all emails).

kolab_policy_greylist and kolab_policy_spf are now just kolab_policy
(with arguments), since they function the same (they are not stateful
like kolab_policy_submission).

Differential Revision: https://git.kolab.org/D5271