Kolabobjects compat mode via config overlays
Configuration overlays are part of the roundcube configuration,
and in our case are dynamically adjusted to remove plugins according to
the deployment configuration.

The API only returns a configuration of which configuration overlays to
apply, with the plugin doing the job of merging the overlay into the
actual configuration.

This allows us to change the configuration from kolab4, without leaking
to much information about individual plugins etc. and while preserving
the capability to have different configurations per deployment, without
having to adjust the kolab4 configuration accordingly. In other words;
it separates concerns better.

Differential Revision: https://git.kolab.org/D5310